% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_jqueryui}
\alias{dependency_jqueryui}
\title{Add a dependency on jQuery UI}
\usage{
dependency_jqueryui(version = "1.12.1")
}
\arguments{
\item{version}{The version of jQuery UI to use}
}
\description{
Add this to the list of dependencies passed to \code{\link{shinywc_ui}}.
}
