% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependency_webcomponentsjs}
\alias{dependency_webcomponentsjs}
\title{Add a dependency on webcomponentsjs polyfill}
\usage{
dependency_webcomponentsjs(version = "2.5.0")
}
\arguments{
\item{version}{The version of webcomponentsjs polyfill to use}
}
\description{
Add a dependency on webcomponentsjs polyfill
}
