% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputknob.R
\name{inputknob}
\alias{inputknob}
\title{Create a \verb{<input-knob>} web component}
\usage{
inputknob(
  id = NULL,
  value = NULL,
  scale = NULL,
  min = NULL,
  max = NULL,
  slot = NULL,
  `slot-back-side` = NULL,
  `css-knob-size` = NULL,
  ...
)
}
\arguments{
\item{id}{ID for the component. If not provided, an ID will be automatically
generated.}

\item{slot}{Content to place in the default (unnamed) slot.}

\item{...}{Any additional HTML attributes to add to the element tag.}
}
\description{
Binding to the \verb{<input-knob>} web component version 1.0.0.
}
