% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputknob.R
\name{inputknob_get_attr}
\alias{inputknob_get_attr}
\title{Retrieve the value of an attribute from the component}
\usage{
inputknob_get_attr(id, attr, cb)
}
\arguments{
\item{cb}{Callback function that gets called when the value is retrieved.
The function has has one argument which holds the value of the attribute.}
}
\description{
Retrieve the value of an attribute from the component
}
